requirements.txt
langchain==0.1.16
chromadb==0.4.24
openai==1.23.6
tiktoken==0.6.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
pandas==2.2.1
streamlit==1.33.0